Organizes page js files into a folder
