+++ src/js/App.js
@@ -1,11 +1,11 @@
 import React, { Component } from 'react';
 
-import Home from './Home';
-import Java from './Java';
-import Unity from './Unity';
-import Web from './Web';
-import Gaming from './Gaming';
-import Music from './Music';
+import Home from './pages/Home';
+import Java from './pages/Java';
+import Unity from './pages/Unity';
+import Web from './pages/Web';
+import Gaming from './pages/Gaming';
+import Music from './pages/Music';
 import AppRouter from './AppRouter';
 
 var currentIndex = 1;
