+++ src/js/pages/Gaming.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 
-import { updateCurrentIndex } from "./App";
+import { updateCurrentIndex } from "../App";
 
 class Gaming extends Component {
