+++ src/js/pages/Music.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 
-import { updateCurrentIndex } from "./App";
+import { updateCurrentIndex } from "../App";
 
 class Music extends Component {
